Introduced sass and css modules and refactored accordingly
