+++ src/App.scss
@@ -1,3 +1,5 @@
+@import './sass/variables.scss';
+
 .App {
   text-align: center;
 }
@@ -14,7 +16,7 @@
 }
 
 .App-header {
-  background-color: #282c34;
+  background-color: $primary-color;
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -25,7 +27,7 @@
 }
 
 .App-link {
-  color: #61dafb;
+  color: $secondary-color
 }
 
 @keyframes App-logo-spin {
