+++ src/components/Navbar/navbar.module.scss
@@ -1,0 +1,39 @@
+@import '../../sass/variables';
+
+header {
+    background: #fff;
+  }
+
+.container {
+    max-width: 1400px;
+    margin: 0 auto;
+    padding: 10px 20px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+
+  header a {
+    color: #333;
+    text-decoration: none;
+  }
+
+
+  nav {
+    display: flex;
+    align-items: center;
+  }
+  nav a {
+    margin-left: 10px;
+  }
+  
+  nav button {
+    background: #fff;
+    color: var(--primary);
+    border: 2px solid var(--primary);
+    padding: 6px 10px;
+    border-radius: 4px;
+    font-family: "Poppins";
+    cursor: pointer;
+    font-size: 1em;
+  }
