+++ src/routes/Home/home.module.scss
@@ -1,0 +1,8 @@
+@import '../../sass/variables';
+
+
+.home {
+    display: grid;
+    grid-template-columns: 3fr 1fr;
+    gap: 100px;
+  }
