+++ src/sass/_variables.scss
@@ -1,0 +1,2 @@
+$primary-color: #282c34;
+$secondary-color: #61dafb;
